Added code that will replace any "/" that might exist in the inputted String from the user with File.separator. This will give it a more OS independent structure.
Added the cmdRemove method. Allowing to remove ALL zits by using the "-a" parameter and/or also specific zits by using the "-z" parameter followed by key names separated by ",".

Renamed some method parameters. To get a more persistent look and read.

Fixed the default "show" command parameter so it is "-n". Which shows the file/folder -names

Fixed the remove command to be more strict on the matching of what was inputted by user.

Fixed the show command to be more strict on the matching of what was inputted by user.